+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+Note: Breaking changes between versions are indicated by "💥".
+
+## Latest
+
+- 💥[Improvement] Replace the `databases` command by `init`.
+
 ## 3.3.11 (2019-06-07)
 
 - [Improvement] Upgrade to ironwood.2
@@ -8,7 +14,7 @@
 - [Bugfix] Fix installing a locally cloned requirement repository
 - [Improvement] Add `--no-cache` option to `images build`
 - [Improvement] Make it possible to configure the notes service hostname
-- [Improvement] Better, more robust MySQL initialization
+- [Improvement] Better, more robust MySQL initialisation
 
 ## 3.3.10 (2019-05-15)
 
@@ -128,7 +134,7 @@
 
 ## 3.0.1 (2019-02-11)
 
-- [Bugfix] fix mysql initialization (#159, #160)
+- [Bugfix] fix mysql initialisation (#159, #160)
 - [Improvement] Better handling of continuous integration
 - [Bugfix] fix `tutor --version` (#156)
 - [Improvement] Absolute settings imports -- 📯 thanks @tonytan4ever!